feat: add safe types for mifare classic block and sectors
