+++ rust/gallagher-rfid-core/src/mifare/mod.rs
@@ -1,0 +1,2 @@
+pub mod application_directory;
+pub mod classic;
